requirements.txt
Django
psycopg2
requests
django-pipeline
django-cprofile-middleware
PyExecJS
